# Speedup arithmetic kernels (up to -25%) / not (-30%)