--- a/arrow-array/src/array/primitive_array.rs
+++ b/arrow-array/src/array/primitive_array.rs
@@ -1029,7 +1029,7 @@ impl<T: ArrowPrimitiveType> PrimitiveArray<T> {
     {
         let nulls = left.logical_nulls();
         let buffer: Vec<_> = (0..left.len())
-            // SAFETY: i is within bounds of the array
+            // SAFETY: i in range 0..left.len()
             .map(|i| op(unsafe { left.value_unchecked(i) }))
             .collect();