# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->


Closes #.

# Rationale for this change
 

<details>

```
add(0)                  time:   [8.2259 µs 8.2342 µs 8.2441 µs]
                        change: [-0.5010% -0.2757% -0.0590%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

add_checked(0)          time:   [7.0879 µs 7.3277 µs 7.6011 µs]
                        change: [-6.7853% -4.9254% -3.0503%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 25 outliers among 100 measurements (25.00%)
  20 (20.00%) low severe
  2 (2.00%) high mild
  3 (3.00%) high severe

add_scalar(0)           time:   [3.6041 µs 3.6297 µs 3.6584 µs]
                        change: [-38.775% -37.843% -37.013%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) high mild
  14 (14.00%) high severe

subtract(0)             time:   [6.4811 µs 6.5582 µs 6.6578 µs]
                        change: [-14.791% -12.619% -10.742%] (p = 0.00 < 0.05)
                        Performance has improved.

subtract_checked(0)     time:   [7.9566 µs 7.9676 µs 7.9805 µs]
                        change: [-4.4506% -4.1728% -3.9161%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe

subtract_scalar(0)      time:   [3.8308 µs 3.9523 µs 4.1161 µs]
                        change: [-23.490% -19.635% -15.749%] (p = 0.00 < 0.05)
                        Performance has improved.

multiply(0)             time:   [6.8732 µs 6.8898 µs 6.9040 µs]
                        change: [-16.540% -16.354% -16.162%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

multiply_checked(0)     time:   [6.7054 µs 6.8740 µs 7.0836 µs]
                        change: [-10.102% -7.8101% -5.9373%] (p = 0.00 < 0.05)
                        Performance has improved.

multiply_scalar(0)      time:   [3.6219 µs 3.6418 µs 3.6607 µs]
                        change: [-37.007% -36.422% -35.888%] (p = 0.00 < 0.05)
                        Performance has improved.


not                     time:   [103.75 ns 103.80 ns 103.86 ns]
                        change: [-28.869% -28.724% -28.584%] (p = 0.00 < 0.05)
                        Performance has improved.

not_sliced              time:   [226.43 ns 226.48 ns 226.52 ns]
                        change: [-13.416% -13.278% -13.154%] (p = 0.00 < 0.05)
                        Performance has improved.

```
</details>

<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please call them out.
-->