@alamb 2025-04-30
> Other than that I think this is still improving the kernels on the use of safety / showing the from_trusted_len_iter is not needed as much.

I agree this PR is an improvement in general. I'll rerun the benchmarks to see if they are repeatable