@alamb 2025-04-30
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_arrow.sh) Running
Linux aal-dev 6.11.0-1013-gcp #13~24.04.1-Ubuntu SMP Wed Apr  2 16:34:16 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing speedup_arith (3f5f977762c73901fb59831aa44cef234aba1dc9) to 07093a49eface9be9208dd427b810abba8d0a755 [diff](https://github.com/apache/arrow-rs/compare/07093a49eface9be9208dd427b810abba8d0a755..3f5f977762c73901fb59831aa44cef234aba1dc9)
BENCH_NAME=boolean_kernels
BENCH_COMMAND=cargo bench --all-features --bench boolean_kernels 
BENCH_FILTER=
BENCH_BRANCH_NAME=speedup_arith
Results will be posted here when complete