@alamb 2025-04-30
🤖: Benchmark completed

<details><summary>Details</summary>
<p>


```
group         main                                   speedup_arith
-----         ----                                   -------------
and           1.00    290.7±0.74ns        ? ?/sec    1.00    291.6±2.83ns        ? ?/sec
and_sliced    1.00   1137.9±1.86ns        ? ?/sec    1.00   1136.3±1.56ns        ? ?/sec
not           1.13    264.3±2.53ns        ? ?/sec    1.00    232.9±0.34ns        ? ?/sec
not_sliced    1.41   1007.1±1.12ns        ? ?/sec    1.00    712.8±0.63ns        ? ?/sec
or            1.07    283.5±0.75ns        ? ?/sec    1.00    265.2±0.30ns        ? ?/sec
or_sliced     1.00   1134.8±1.18ns        ? ?/sec    1.00   1140.2±1.55ns        ? ?/sec
```


</p>
</details>